cortex/requirements.txt
crewai
crewai-tools
langchain
pydantic
python-dotenv
openai
